+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 ## Unreleased
 
+## 0.0.12 (2025-05-16)
+
 ### Changed
 
-- Respond to cancel for exec builds. (<https://github.com/pulumi/pulumi-docker-build/pull/522>)
+- Upgraded pulumi-go-provider to v1.0.0-rc2.
 
 ### Fixed
 
-- Builds now respect cancellation. (https://github.com/pulumi/pulumi-docker-build/issues/533)
+- Builds now respect cancellation. (https://github.com/pulumi/pulumi-docker-build/issues/533, https://github.com/pulumi/pulumi-docker-build/pull/522)
 
 ## 0.0.11 (2025-04-11)